restructure constructor code
