+++ python/hsfs/constructor/on_demand_feature_group_alias.py
@@ -21,8 +21,10 @@ from hsfs import on_demand_feature_group as on_demand_fg
 
 class OnDemandFeatureGroupAlias:
     def __init__(self, on_demand_feature_group, alias):
-        self._on_demand_feature_group = on_demand_fg.OnDemandFeatureGroup.from_response_json(
-            on_demand_feature_group
+        self._on_demand_feature_group = (
+            on_demand_fg.OnDemandFeatureGroup.from_response_json(
+                on_demand_feature_group
+            )
         )
         self._alias = alias
